refactor(core): split feats into sibling modules with facade
Move text/apply/requirements out of feats.py; keep stable core.feats
re-exports for callers.
